+++ src/view/secondary-bar/test/calendar-sort-criteria.test.ts
@@ -1,9 +1,12 @@
 import { faker } from '@faker-js/faker';
 import { FOLDERS } from '@zextras/carbonio-ui-commons';
 
-import { getCalendarSortCriteria } from './calendar-sort-criteria';
-import { generateFolder, generateFolderLink } from '../../__test__/mocks/folders/folders-generator';
-import { getMocksContext } from '../../__test__/mocks/utils/mocks-context';
+import {
+	generateFolder,
+	generateFolderLink
+} from '../../../__test__/mocks/folders/folders-generator';
+import { getMocksContext } from '../../../__test__/mocks/utils/mocks-context';
+import { getCalendarSortCriteria } from '../calendar-sort-criteria';
 
 /*
  * SPDX-FileCopyrightText: 2025 Zextras <https://www.zextras.com>
