+++ src/view/secondary-bar/test/use-accordion-item-open-status-storage.test.ts
@@ -4,9 +4,9 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-import { useAccordionItemOpenStatusStorage } from './use-accordion-item-open-status-storage';
-import { useLocalStorage } from '../../__test__/mocks/carbonio-shell-ui/carbonio-shell-ui';
-import { setupHook } from '../../__test__/test-setup';
+import { useLocalStorage } from '../../../__test__/mocks/carbonio-shell-ui/carbonio-shell-ui';
+import { setupHook } from '../../../__test__/test-setup';
+import { useAccordionItemOpenStatusStorage } from '../use-accordion-item-open-status-storage';
 
 describe('useAccordionItemOpenStatusStorage', () => {
 	it('should return an object with 2 functions to get and set the open status', () => {
